post

+++ app/Http/Controllers/TasksController.php
@@ -16,12 +16,24 @@ class TasksController extends Controller
     public function index()
     {
         // タスク一覧を取得
-        $tasks = Task::all();
-
+        //$tasks = Task::all();
+        
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+        
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
         //タスク一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        return view('welcome',$data);
+
     }
 
     /**
@@ -54,10 +66,22 @@ class TasksController extends Controller
         ]);
         
         // タスクを作成
-        $task = new Task;
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
+        //$task = new Task;
+       // $task->status = $request->status;    // 追加
+       // $task->content = $request->content;
+      
+        //$task->save();
+        
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+            
+            
+        ]);
+
+        
+        
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -73,9 +97,16 @@ class TasksController extends Controller
     {
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
+        
+        // 関係するモデルの件数をロード
+        //$task->loadRelationshipCounts();
 
+        // ユーザの投稿一覧を作成日時の降順で取得
+        //$tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+        
         // タスク詳細ビューでそれを表示
         return view('tasks.show', [
+            //'user' => $user,
             'task' => $task,
             ]);
     }
@@ -133,9 +164,17 @@ class TasksController extends Controller
     {
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
+        
+        //dd($task);
         // タスクを削除
-        $task->delete();
-
+        //$task->delete();
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+        
+        
         // トップページへリダイレクトさせる
         return redirect('/');
     }
